refactor: ShortestPathFactory의 역할을 PathType로 통합
