+++ README.md
@@ -30,5 +30,5 @@
 ### 작업목록
 - [x] Cucumber 컨밴션 통일
 - [x] ShortestPath 에서 Station 정보 추출
-- [ ] ShortestPathFactory 존재 의미 고민
+- [x] ShortestPathFactory 존재 의미 고민
 - [ ] 경로 조회 결과에 요금 정보 추가
