+++ src/main/java/nextstep/path/application/PathService.java
@@ -2,9 +2,8 @@ package nextstep.path.application;
 
 import nextstep.line.domain.SectionRepository;
 import nextstep.path.application.dto.PathsResponse;
-import nextstep.path.domain.ShortestPath;
-import nextstep.path.domain.ShortestPathFactory;
 import nextstep.path.domain.PathType;
+import nextstep.path.domain.ShortestPath;
 import nextstep.path.ui.exception.SameSourceAndTargetException;
 import nextstep.station.application.StationService;
 import nextstep.station.application.dto.StationResponse;
@@ -15,7 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static nextstep.path.domain.PathType.*;
+import static nextstep.path.domain.PathType.DISTANCE;
 
 @Service
 @Transactional(readOnly = true)
@@ -49,7 +48,7 @@ public class PathService {
 
     private ShortestPath createShortestPath(Long source, Long target, PathType pathType) {
         validateSameSourceAndTarget(source, target);
-        return ShortestPathFactory.createShortestPath(sectionRepository.findAll(), pathType);
+        return pathType.createShortestPath(sectionRepository.findAll());
     }
 
     private static void validateSameSourceAndTarget(Long source, Long target) {
